* VStructure, VDefault: (caseSatisfyRequirementUsage): Refactored it   by moving it from VDefault to VStructure.
